partial update to fix the test coverage for the factory
