+++ src/dodal/beamlines/i22.py
@@ -5,6 +5,7 @@ from ophyd_async.epics.adpilatus import PilatusDetector
 from ophyd_async.fastcs.panda import HDFPanda
 
 from dodal.common.beamlines.beamline_utils import (
+    BeamlinePrefix,
     get_path_provider,
     set_path_provider,
 )
@@ -25,7 +26,7 @@ from dodal.devices.synchrotron import Synchrotron
 from dodal.devices.tetramm import TetrammDetector
 from dodal.devices.undulator import Undulator
 from dodal.log import set_beamline as set_log_beamline
-from dodal.utils import BeamlinePrefix, get_beamline_name
+from dodal.utils import get_beamline_name
 
 BL = get_beamline_name("i22")
 set_log_beamline(BL)
