+++ src/dodal/beamlines/p38.py
@@ -4,9 +4,11 @@ from ophyd_async.epics.adaravis import AravisDetector
 from ophyd_async.fastcs.panda import HDFPanda
 
 from dodal.common.beamlines.beamline_utils import (
+    BeamlinePrefix,
     device_instantiation,
     get_path_provider,
     set_path_provider,
+    skip_device,
 )
 from dodal.common.beamlines.beamline_utils import set_beamline as set_utils_beamline
 from dodal.common.beamlines.device_helpers import numbered_slits
@@ -19,7 +21,7 @@ from dodal.devices.slits import Slits
 from dodal.devices.tetramm import TetrammDetector
 from dodal.devices.undulator import Undulator
 from dodal.log import set_beamline as set_log_beamline
-from dodal.utils import BeamlinePrefix, get_beamline_name, skip_device
+from dodal.utils import get_beamline_name
 
 BL = get_beamline_name("p38")
 set_log_beamline(BL)
