+++ src/dodal/common/beamlines/device_helpers.py
@@ -1,6 +1,5 @@
-from dodal.common.beamlines.beamline_utils import device_instantiation
+from dodal.common.beamlines.beamline_utils import device_instantiation, skip_device
 from dodal.devices.slits import Slits
-from dodal.utils import skip_device
 
 
 @skip_device()
